Use hexagonal architecture
